Engraving: move chord symbol realization from Score into RealizeChordSymbols
Moves cmdRealizeChordSymbols (and its file-static setChord helper) out of
Score into a new static RealizeChordSymbols class.
