+++ src/engraving/editing/realizechordsymbols.h
@@ -1,0 +1,43 @@
+/*
+ * SPDX-License-Identifier: GPL-3.0-only
+ * MuseScore-Studio-CLA-applies
+ *
+ * MuseScore Studio
+ * Music Composition & Notation
+ *
+ * Copyright (C) 2026 MuseScore Limited
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+
+namespace mu::engraving {
+class Score;
+class Transaction;
+
+enum class Voicing : signed char;
+enum class HDuration : signed char;
+
+class RealizeChordSymbols
+{
+public:
+    // Realize selected chord symbols into notes on the staff.
+    //
+    // If a voicing and duration type are specified, the harmony voicing settings will be
+    // overridden by the passed parameters. Otherwise, the settings set on the harmony object
+    // will be used.
+    static void realizeChordSymbols(Transaction& tx, Score* score, bool literal = true, Voicing voicing = Voicing(-1),
+                                    HDuration durationType = HDuration(-1));
+};
+}
